requirements.txt
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
xmltodict==0.13.0
tensorflow==2.13.0
matplotlib==3.7.2
scikit-learn==1.3.0
tqdm==4.65.0
seaborn==0.12.2
imgaug==0.4.0
albumentations==1.3.1 